backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
argon2-cffi==25.1.0
redis==5.2.1
httpx==0.28.1
python-multipart==0.0.20
groq==0.11.0

# Test bağımlılıkları
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.22.1
hypothesis==6.122.3
pytest-mock==3.14.0
factory-boy==3.3.1
ruff==0.8.4
